am: tuple_foreach を am/util.h に移動した
